# server/openapi.yaml
openapi: 3.1.0
info:
  title: RelationBoat API
  version: 0.1.0
servers:
  - url: https://relationboat.kro.kr/api/v1
  - url: http://localhost:3000/api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Person:
      type: object
      required: [id, folderId, name]
      properties:
        id: { type: string }
        folderId: { type: string }
        name: { type: string }
        phone: { type: [string, 'null'] }
        memo: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
security:
  - bearerAuth: []
paths:
...
  /folders/{folderId}/people:
    parameters:
      - in: path
        name: folderId
        required: true
        schema: { type: string }
    get:
      summary: List people in folder scope only.
      responses:
        '200':
          description: People list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Person' }
    post:
      summary: Create person.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1 }
                phone: { type: [string, 'null'] }
                memo: { type: [string, 'null'] }
      responses:
        '201': { description: Created }
  /folders/{folderId}/people/{personId}:
    parameters:
      - in: path
        name: folderId
        required: true
        schema: { type: string }
      - in: path
        name: personId
        required: true
        schema: { type: string }
    get:
      summary: Get person inside folder scope.
      responses:
        '200': { description: Person detail }
    patch:
      summary: Update person.
      responses:
        '200': { description: Updated }
    delete:
      summary: Delete person.
      responses:
        '204': { description: Deleted }
...